Make large map generation more interesting
Large maps are big, dry, and boring. When you generate a world with > 5 players it now adds a big ocean strip on the ... equator?
does nothing for <5 players

Also make islands more interesting. Previously islands could only generate with forests, oil, and xil. Now they have a chance to have coal or iron in several configurations.

The most powerful weapon in the game is a ship, so it's nice to have some ocean to play with